refactor TransportServer controller
